requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.0
    # via -r requirements.in
cffi==1.15.0
    # via bcrypt
click==8.1.2
    # via uvicorn
dnspython==2.2.1
    # via email-validator
email-validator==1.1.3
    # via pydantic
fastapi==0.75.2
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
passlib==1.7.4
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic[email]==1.9.0
    # via
    #   -r requirements.in
    #   fastapi
pyjwt==2.3.0
    # via -r requirements.in
pymongo==4.1.1
    # via -r requirements.in
python-decouple==3.6
    # via -r requirements.in
six==1.16.0
    # via bcrypt
sniffio==1.2.0
    # via anyio
starlette==0.17.1
    # via fastapi
typing-extensions==4.2.0
    # via pydantic
uvicorn==0.17.6
    # via -r requirements.in
